Fixed text input answer height Fixed answer list dividers Display first term and conditions then make the register api call added some content for about page
