+++ survey/src/main/assets/about.html
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset='utf-8'>
+    <meta content='width=device-width' name='viewport'>
+</head>
+<body>
+<strong>About</strong>
+<p>
+    Welcome to About, your number one source for all things [product]. We're dedicated to giving you
+    the very best of [product], with a focus on [store characteristic 1], [store characteristic 2],
+    [store characteristic 3].
+</p>
+<p>
+    Founded in [year] by [founder name], About has come a long way from its beginnings in [starting
+    location]. When [founder name] first started out, [his/her/their] passion for [brand message -
+    e.g. "eco-friendly cleaning products"] drove them to [action: quit day job, do tons of research,
+    etc.] so that About can offer you [competitive differentiator - e.g. "the world's most advanced
+    toothbrush"]. We now serve customers all over [place - town, country, the world], and are
+    thrilled that we're able to turn our passion into [my/our] own website.
+</p>
+<p>
+    [I/we] hope you enjoy [my/our] products as much as [I/we] enjoy offering them to you. If you
+    have any questions or comments, please don't hesitate to contact [me/us].
+</p>
+<div><p>
+    Sincerely,
+
+    [founder name]
+</p>
+</div>
+</body>
+</html>
