+++ survey/src/main/java/pandemic/response/framework/AboutActivity.kt
@@ -1,6 +1,7 @@
 package pandemic.response.framework
 
 import android.os.Bundle
+import androidx.core.text.HtmlCompat
 import pandemic.response.framework.databinding.AboutActivityBinding
 
 class AboutActivity : BaseActivity() {
@@ -16,5 +17,12 @@ class AboutActivity : BaseActivity() {
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
         supportActionBar?.setDisplayShowHomeEnabled(true)
         title = getString(R.string.about_us)
+        binding.aboutContent.text =
+            HtmlCompat.fromHtml(getAboutContent(), HtmlCompat.FROM_HTML_MODE_LEGACY)
     }
+
+    private fun getAboutContent(): String =
+        assets.open("about.html").bufferedReader().use {
+            it.readText()
+        }
 }
