+++ survey/src/main/res/layout/answer_multiple_choice_question.xml
@@ -4,4 +4,5 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:button="@null"
-    android:drawableEnd="?android:attr/listChoiceIndicatorSingle" />
+    android:drawableEnd="?android:attr/listChoiceIndicatorSingle"
+    android:textSize="16sp" />
