+++ survey/src/main/res/layout/question_boolean.xml
@@ -21,7 +21,9 @@
         android:id="@+id/singleAnswersGroup"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_margin="@dimen/margin_xlarge">
+        android:layout_margin="@dimen/margin_xlarge"
+        android:divider="@color/neutral_one"
+        android:showDividers="middle">
 
         <RadioButton
             android:id="@+id/radioYes"
