+++ survey/src/main/res/layout/question_text.xml
@@ -36,9 +36,9 @@
         <com.google.android.material.textfield.TextInputEditText
             android:layout_width="match_parent"
             android:layout_height="match_parent"
+            android:layout_weight="1"
             android:gravity="start|top"
-            android:inputType="textMultiLine"
-            android:lines="2" />
+            android:inputType="textMultiLine" />
 
     </com.google.android.material.textfield.TextInputLayout>
